apps/duplicate-checker/requirements.txt
# Work Version: python 3.13.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
faiss-cpu==1.10.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
ImageHash==4.3.2
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
python-multipart==0.0.20
PyWavelets==1.8.0
scipy==1.15.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
